+++ src/main/java/com/moyo/backend/domain/batch/ranking/dto/RankingPreflight.java
@@ -1,4 +1,4 @@
-package com.moyo.backend.batch.ranking.dto;
+package com.moyo.backend.domain.batch.ranking.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
