+++ src/main/java/com/moyo/backend/domain/batch/ranking/processor/RankingCalculatorParameters.java
@@ -1,4 +1,4 @@
-package com.moyo.backend.batch.ranking.processor;
+package com.moyo.backend.domain.batch.ranking.processor;
 
 public record RankingCalculatorParameters(
 	int stars,
